Rename files and move index.ts to src
