+++ src/services/heartbeat.service.ts
@@ -1,0 +1,11 @@
+import { HeartbeatModel } from "../models/heartbeat.model";
+import { HeartbeatRepository } from "../repositories/heartbeat.repository";
+
+class HeartbeatService {
+  public getBeat(): HeartbeatModel {
+    const heartbeatRepositoryInstance = new HeartbeatRepository();
+    return heartbeatRepositoryInstance.getBeat();
+  }
+}
+
+export const heartbeatService = new HeartbeatService();
